requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --strip-extras requirements-dev.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via web3
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.21.0
    # via tortoise-orm
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via
    #   -r /Users/0xfbravo/Developer/mb/requirements.in
    #   asyncpg-stubs
asyncpg-stubs==0.30.2
    # via -r requirements-dev.in
attrs==25.3.0
    # via aiohttp
authlib==1.6.0
    # via safety
bitarray==3.4.3
    # via eth-account
black==25.1.0
    # via -r requirements-dev.in
cached-property==2.0.1
    # via py-evm
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
ckzg==2.1.1
    # via
    #   eth-account
    #   py-evm
click==8.0.2
    # via
    #   -r /Users/0xfbravo/Developer/mb/requirements.in
    #   black
    #   nltk
    #   safety
    #   typer
    #   uvicorn
coverage==7.9.2
    # via pytest-cov
cryptography==45.0.5
    # via authlib
cytoolz==1.0.1
    # via eth-utils
dnspython==2.7.0
    # via email-validator
dparse==0.6.4
    # via
    #   safety
    #   safety-schemas
email-validator==2.2.0
    # via -r /Users/0xfbravo/Developer/mb/requirements.in
eth-abi==5.2.0
    # via
    #   eth-account
    #   eth-tester
    #   web3
eth-account==0.13.7
    # via
    #   eth-tester
    #   web3
eth-bloom==3.1.0
    # via py-evm
eth-hash==0.7.1
    # via
    #   eth-bloom
    #   eth-tester
    #   eth-utils
    #   trie
    #   web3
eth-keyfile==0.8.1
    # via eth-account
eth-keys==0.7.0
    # via
    #   eth-account
    #   eth-keyfile
    #   eth-tester
    #   py-evm
eth-rlp==2.2.0
    # via eth-account
eth-tester==0.13.0b1
    # via web3
eth-typing==5.2.1
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   py-ecc
    #   py-evm
    #   web3
eth-utils==5.3.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   eth-tester
    #   py-ecc
    #   py-evm
    #   rlp
    #   trie
    #   web3
fastapi==0.115.14
    # via -r /Users/0xfbravo/Developer/mb/requirements.in
filelock==3.16.1
    # via safety
flake8==7.3.0
    # via -r requirements-dev.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hexbytes==1.3.1
    # via
    #   eth-account
    #   eth-rlp
    #   trie
    #   web3
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements-dev.in
    #   safety
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
iso8601==2.1.0
    # via tortoise-orm
isort==6.0.1
    # via -r requirements-dev.in
jinja2==3.1.6
    # via safety
joblib==1.5.1
    # via nltk
lefthook==1.11.16
    # via -r requirements-dev.in
loguru==0.7.3
    # via -r /Users/0xfbravo/Developer/mb/requirements.in
lru-dict==1.3.0
    # via py-evm
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==4.0.0
    # via safety
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
mypy==1.16.1
    # via -r requirements-dev.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
nltk==3.9.1
    # via safety
packaging==25.0
    # via
    #   black
    #   dparse
    #   pytest
    #   safety
    #   safety-schemas
parsimonious==0.10.0
    # via eth-abi
pathspec==0.12.1
    # via
    #   black
    #   mypy
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psutil==6.1.1
    # via safety
psycopg==3.2.9
    # via tortoise-orm
psycopg-binary==3.2.9
    # via psycopg
psycopg-pool==3.2.6
    # via psycopg
py-ecc==8.0.0
    # via py-evm
py-evm==0.12.1b1
    # via eth-tester
py-geth==6.0.0
    # via web3
pycodestyle==2.14.0
    # via flake8
pycparser==2.22
    # via cffi
pycryptodome==3.23.0
    # via
    #   eth-hash
    #   eth-keyfile
pydantic==2.9.2
    # via
    #   -r /Users/0xfbravo/Developer/mb/requirements.in
    #   eth-account
    #   eth-utils
    #   fastapi
    #   py-geth
    #   safety
    #   safety-schemas
    #   web3
pydantic-core==2.23.4
    # via pydantic
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via
    #   pytest
    #   rich
pypika-tortoise==0.6.1
    # via tortoise-orm
pytest==8.4.1
    # via
    #   -r requirements-dev.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==1.0.0
    # via -r requirements-dev.in
pytest-cov==6.2.1
    # via -r requirements-dev.in
python-dotenv==1.1.1
    # via
    #   -r /Users/0xfbravo/Developer/mb/requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via -r /Users/0xfbravo/Developer/mb/requirements.in
pytz==2025.2
    # via tortoise-orm
pyunormalize==16.0.0
    # via web3
pyyaml==6.0.2
    # via
    #   -r /Users/0xfbravo/Developer/mb/requirements.in
    #   uvicorn
regex==2024.11.6
    # via
    #   nltk
    #   parsimonious
requests==2.32.4
    # via
    #   -r /Users/0xfbravo/Developer/mb/requirements.in
    #   py-geth
    #   safety
    #   web3
rich==14.0.0
    # via typer
rlp==4.1.0
    # via
    #   eth-account
    #   eth-rlp
    #   eth-tester
    #   py-evm
    #   trie
ruamel-yaml==0.18.14
    # via
    #   safety
    #   safety-schemas
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
safe-pysha3==1.0.5
    # via eth-hash
safety==3.5.2
    # via -r requirements-dev.in
safety-schemas==0.0.14
    # via safety
semantic-version==2.10.0
    # via
    #   eth-tester
    #   py-geth
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sortedcontainers==2.4.0
    # via trie
starlette==0.46.2
    # via fastapi
tenacity==9.1.2
    # via safety
tomlkit==0.13.3
    # via safety
toolz==1.0.0
    # via cytoolz
tortoise-orm==0.25.1
    # via -r /Users/0xfbravo/Developer/mb/requirements.in
tqdm==4.67.1
    # via nltk
trie==3.1.0
    # via py-evm
typer==0.16.0
    # via safety
types-pyyaml==6.0.12.20250516
    # via -r requirements-dev.in
types-requests==2.32.4.20250611
    # via
    #   py-geth
    #   web3
typing-extensions==4.14.1
    # via
    #   aiosqlite
    #   asyncpg-stubs
    #   eth-typing
    #   fastapi
    #   mypy
    #   psycopg-pool
    #   py-geth
    #   pydantic
    #   pydantic-core
    #   safety
    #   safety-schemas
    #   typer
    #   web3
urllib3==2.5.0
    # via
    #   requests
    #   types-requests
uvicorn==0.35.0
    # via -r /Users/0xfbravo/Developer/mb/requirements.in
uvloop==0.21.0
    # via uvicorn
watchdog==6.0.0
    # via -r requirements-dev.in
watchfiles==1.1.0
    # via uvicorn
web3==7.12.0
    # via
    #   -r /Users/0xfbravo/Developer/mb/requirements.in
    #   -r requirements-dev.in
websockets==15.0.1
    # via
    #   uvicorn
    #   web3
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
